update finalo/karaoke
doesn't work but is an example of nested plugins
