--- sigma/plugins/finalo/__init__.py
@@ -1,11 +1,0 @@
-from .karaoke import VoiceChangeDetection
-from .karaoke import Control
-from .selfrole import SelfRole
-
-__all__ = [
-    'VoiceChangeDetection',
-    'Control',
-    'SelfRole'
-]
-
-pluglist = __all__
