+++ sigma/plugins/finalo/karaoke/forcemute.py
@@ -1,0 +1,5 @@
+async def forcemute(cmd, message, args):
+    if checkPermissions(message.author):
+        print('Force muting')
+        await mutekaraokechannel()
+        await cmd.reply("Iterated through channel")
