+++ sigma/plugins/finalo/karaoke/plugin.yml
@@ -1,0 +1,6 @@
+name: Karaoke
+
+categories:
+  - finalo
+
+enabled: true
