+++ sigma/plugins/finalo/plugin.yml
@@ -1,0 +1,6 @@
+name: Finalo
+
+categories:
+  - finalo
+
+enabled: true
